Fix parsing of QueryResult responses after recent API changes
Starting some days ago Alma added a namespace to the "rowset" tag in
the QueryResult response. Adjust our xpath to ignore namespaces there
to work with the old and the new format, and move the parsing in
a central place.

This fixes all APIs using the analytics API internally to not return empty
responses after the Alma update.