Merge in the base org connector and remove the connector interface
For historic reasons we had a connector to abstract away campusonline,
the the connector was moved to just be a base connector. It's not really
work abstracting that away and it's fine to hard depend on the base org,
so merge in the connector + the config.

Move the dummy implementations into the test namespace.
It's no longer used as a fallback implementation in case there is
no connector, so move it to the tests where it is still used.