sublibrary: deprecate libraryManager parameter and implement GET
Since the last changes to the API it's no longer possible to pass
a user ID that is not your own to the the collection GET, so might as well
default to the current user and make libraryManager optional and
deprecated.

While at it also implement the item GET using the same logic.